Fix bug of incorrect message size causing "Unknown message error"
Read timeout was set to 1 second in order to be able of processing Notifications
after each second. However, this has proved to be prone to errors. Specifically,
it seems like while reading the header of a message, timeout was triggered after
read a few bytes causing an error in the second attempt because of the already
consumed bytes.

This commit fix this by removing this timeout for reads. Reads are no performed
in a dedicated go routine and packets read are then send through a channel in order
to get it processed by the session.