Support completed for event cancelling
This implementation does some tricks. A cancelled event is one that has
event state CANCELLED and inbox_position field set to the time the event
was cancelled. This implies that a cancelled event will continue being
sent to client as normal, whenever its start date is after current date.

So far, this behaviour is prefered because if cancelled notification isn't
sucessfully delivered to client, events will be in sync next time client
request recent events. If the client request recent events after start_date
of the cancelled event, then client and server will be out of sync, i.e, client
will have the event as finished, but server as cancelled. Anyway, this
do not suppose a different for client.

This behaviour will be changed when client events history is synced with
the server in order to do not send cancelled events as part of recent
events.