Change event manager architecture
This commit is the first one of a series of commits to migrate old
events architecture to a new one. Previous architecture stored events
in events_by_user table in Cassandra where user_id is the partition key
and start date is the cluster key. Because of this, when an event start date
is modified, it has to be deleted and reinserted in Cassandra. Moreover,
this operation has to be repeated for each participant in the event. So, changes
in event dates have a cost of N*(delete+insert) operations. This behaviour
is not desirable.

This new architecture removes events_by_user table and introduces a new one
called events_timeline. The event timeline contains all of the events created
by users in a single time line backed up by Cassandra. This way it is now more
easy than ever to loop all of the events in an ordered way. Because users
still need to retrieve their own events and it would be unfeasable to look for
them in the whole timeline each time it is required, user events are managed
in RAM. Not all events of a user are managed in RAM but only those that have not
finished yet.

Due to events RAM management, it is needed a mechanism to expire events when they
finish. A goroutine that is scheduled each minute is responsible for this. This
routine will check finished events since last time (while running and at start up)
and archive them for each users. This archive is hold by events_history_by_user
table in Cassandra.

Modify server to use latest changes