Make cancelled events reacheable in time line
When a event was cancelled it was moved in timeline from event end date
position to current date position, where end date > current date. When current
Android and iOS apps receive a cancelled event notification, they retrieve
the recent events. Due to the event being moved in timeline, it was no longer
reacheable as part of recent events.

This commit fixes this behaviour not moving cancelled events in timeline so
that clients can retrieve them and have a chance to get the last event state.